Chore: Reorganizing tests to align with project structure
