+++ tests/test-load.js
@@ -3,8 +3,6 @@
  * @author Nicholas C. Zakas
  */
 
-/* eslint no-undefined:0 */
-
 "use strict";
 
 var assert = require("assert");
@@ -12,6 +10,6 @@ var assert = require("assert");
 describe("eslint generator", function() {
     it("can be imported without blowing up", function() {
         var app = require("../rule");
-        assert(app !== undefined);
+        assert.ok(app);
     });
 });
